requirements.txt
mcp>=1.0
httpx>=0.27
pydantic>=2.0
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=0.24
bedrock-agentcore-starter-toolkit
bedrock-agentcore
uvicorn>=0.30
